# Conflicts:
#	FantasySurvivor/Assets/Resources/Configs/DataCharacterConfig.txt
#	FantasySurvivor/Assets/_App/Prefabs/System/--GameRunner.prefab
#	FantasySurvivor/Assets/_App/Scenes/scn_Loading.unity
#	FantasySurvivor/Assets/_App/Scripts/Configs/DataCharacterConfig.cs
#	FantasySurvivor/Assets/_App/Scripts/Controllers/CharacterController.cs
#	FantasySurvivor/Assets/_App/Scripts/Controllers/GameController.cs
#	FantasySurvivor/Assets/_App/Scripts/Models/CharacterModel.cs
#	FantasySurvivor/Assets/_App/Scripts/Stat/CharacterUpdateStat.cs
#	FantasySurvivor/Assets/_App/Scripts/Stat/StatModel.cs
#	FantasySurvivor/Assets/_App/Scripts/Views/InGame/Character/Character.cs
#	FantasySurvivor/Assets/_App/Scripts/Views/InGame/Skills/PoisonBullet.cs

+++ FantasySurvivor/Assets/_App/Scripts/Models/CharacterModel.cs
@@ -8,7 +8,7 @@ using UnityEngine.U2D;
 public class CharacterModel : Model<GameApp>
 {
 	public static EventTypeBase dataChangedEvent = new EventTypeBase(nameof(CharacterModel) + ".dataChanged");
-	public CharacterModel(float maxHp, float moveSpeed, float attackDamage, float itemAttractionRange, float attackRange, int armor, float regen, float shield) : base(dataChangedEvent)
+	public CharacterModel(float maxHp, float moveSpeed, float attackDamage, float itemAttractionRange, float attackRange, int armor, float regen, int shield) : base(dataChangedEvent)
 	{
 		this.currentHealthPoint = maxHp;
 		this.maxHealthPoint = maxHp;
@@ -42,8 +42,8 @@ public class CharacterModel : Model<GameApp>
 	
 	private float _regen;
 
-    private float _shield;
-    public float shield
+    private int _shield;
+    public int shield
     {
         get => _shield;
         set
@@ -55,13 +55,12 @@ public class CharacterModel : Model<GameApp>
     }
     public float moveSpeed
     {
-        get => _moveSpeed;
-        set
-        {
-            if (moveSpeed.Equals(value)) return;
-            _moveSpeed = value;
-            RaiseDataChanged(nameof(moveSpeed));
-        }
+	    get => _moveSpeed;
+	    set {
+		    if(moveSpeed.Equals(value)) return;
+		    _moveSpeed = value;
+		    RaiseDataChanged(nameof(moveSpeed));
+	    }
     }
 
     public float currentHealthPoint
